requirements.txt
streamlit
transformers
diffusers
torch
gtts
Pillow
requests